Fix the copy ctors/assign ops so they're more consistent
Generally copy assign is implementable as copy-and-swap. This
implements it that way. I'm not super stoked about the fact that the
copy ctor is implemented in terms of a free-function that operates on
objects, since the "*this" object isn't technically well-formed
yet. Then again, this is no worse than the old code, so overall I'm
claiming a win.