Add assertion that LAPACK is found. (#131)
This is hopefully a temporary measure. However, we have, historically,
recommended OpenBLAS as a widely-supported, high-quality BLAS
implementation that also supports the LAPACK API (assuming it's been
built to do so), and many vendors supply an (allegedly) optimized
implementation for their platforms. That said, if your workloads are
primarily GPU-based, a basic installation of Netlib BLAS/LAPACK is
sufficient to fulfill the API requirement and should cost little in
the way of net application performance...